loot manager to manage how loot is generated, temp use a different normal table for bronze rare chest, made mean a float,

+++ src/main/java/teamport/aether/helper/AetherMathHelper.java
@@ -6,14 +6,14 @@ public class AetherMathHelper {
     /**
      * @implNote Exponential can return any value between [0, INF) and as such this function caps it.
      * */
-    public static int invertedExponentialCapped(Random random, int mean, int cap) {
+    public static int invertedExponentialCapped(Random random, float mean, int cap) {
         return Math.min((int)Math.floor(invertedExponential(random, mean)), cap);
     }
 
     /**
      * @implNote Inverse Exponential function where the expected value is the parameter mean.
      * */
-    public static double invertedExponential(Random random, int mean) {
+    public static double invertedExponential(Random random, float mean) {
         return nextExponential(random) * mean;
     }
